drm: Reduce latency while processing atomic ioctls
Unfortunately, our display driver must occasionally sleep inside
start_atomic(), which causes the CPU running the ioctl to schedule and
potentially idle. Depending on how deeply the running CPU idles, there
can be quite a bit of latency added to processing the "atomic" ioctl as
a result, which hurts display rendering latency.

We can mitigate this effect somewhat by forcing the CPU running the
ioctl to stick to shallow idle states. This is done by forcing the task
running the ioctl to stick to its current CPU and then using PM QoS to
keep said CPU from entering idle states that take more than 100 us to
exit.

We can also further reduce the rendering latency by elevating the
priority of the task running the ioctl, so that after it goes to sleep,
it will wake up sooner when the wait inside start_atomic() is over.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>